feat: Extended the value for `lis_result_sourcedid` to be larger than 256 character limit.
Used a sidecar Django model `GradedAssignmentExtra` to add a larger field character limit > 256 characters.

One of our institutions `NorthArkansasCollege` was sending in a `lis_result_sourcedid` value of 640 characters which is greater than the platform defined VARCHAR(256) limit. With this plugin change, we're able to extend the limit using the sidecar model field above.

Also, to ensure that the business logic and testing still call the same `GradedAssignment.lis_result_sourcedid` property, a `_SourcedidProxy` in conjunction with a `_sync_sidecar` method that listens on post_save of the `GradedAssignment` ORM instance was implemented. This proxy handles both getter and setter functions for this property value, setting the `extra.lis_result_sourcedid_long` value and keeping the original `GradedAssignment.lis_result_sourcedid` truncated from the POST value sent to a limit of 256 characters.